consistent json names, user ansible scripts
Telling go serialization to use lowercase names to be consistent with
what is expected outside of goland.

Now executing user provided ansible scripts according to the following
convention. If an ansible script in the model directory exists that has
the name <node>.yml, that script will be executed in the configuration
stage for <node> in the model topology.