support for freebsd guests
tested with freebsd 11. Needed to add support to guests for apic and
acpi as well as change the disk mounting bus target to virtio with a
vda device as opposed to the old sda:sata emulation model.

updated the 3bed model to use freebsd on some nodes. Nodes are coming up
and grabbing addresses from dhcp.