3bed topology coming up with deter config
defined new format for a model directory this is

<model>
|-- model.js
|-- /config
|   |-- [node].yml
|   |-- /files
|       |-- /[node]
|           |--  *.conf
|-- /pre-config
|   |-- run

each model has a top level config, files and pre-config directory in
addition to a json model file 'model.js'.

The config directory contains the ansible configs for nodes in the
experiment.  The convention is that the name of the ansible file is
<node>.yml. Any file with this naming convention under config will be
picked up by raven at config time.

The config/files directory contains configuration files for the nodes.
The convention is that each node has its own directory. However, this
convention does not actually result in any action on ravens part. It's
just a convention.

The preconfig directory, if present, contains an executable script
named 'run'. This is a shell script that is executed by raven just
before configuration takes place. At this time a single environment
variable called TOPOJSON is provided to that script. The script can use
that to call other programs that ingest and use the annotated topology
to create topology specific configuration files.