rvn-ansible real time output, more setup, netboot
some ansible commands take quite a while to go, specifically things that
pull down big files from the internets and such. now the rvn-ansible
command spits out the ansible output as it comes from ansible.

updated the first-user script to be smart about adding users and groups
as mkgroup takes 47 years to time out if the group already exists.

added configuration for the walrus

added the capability for nodes to netboot in the libvirt setup

moved ansible configs for n* nodes out as they are not netboot. also
added additional logic to the base level configuration stage so that it
does not try to configure a netboot node.