starting rvn command line utility, +reboot control
started a command line utility for raven in rvn-cli. The design of this
command line utility is to go through the http api for everything, so it
really shadows the functionality of the web interface.

added a reboot command to the api. this command is accessible through
the new rvn cli utility.