configure per node, wipe command
There is now the option to run `configure` on a per node basis instead
of all nodes. `configure` without any node arguments retains the
existing behavior. `configure` with a list of node arguments only
configures the specified nodes.

There is a new command called `wipe`. `wipe` destroys a virtual machine
including its virtual disk and then recreates the machine with a fresh
virtual disk. Configure is not automatically rerun, so if the machine
was configured previously `configure` will need to be rerun on the
machine to bring it back to the configured state. Wipe is asynchronous
with respect to the machine actually coming up. Use `pingwait` inbetween
a `wipe` and `configure` or any other subsequent actions.