raven: non-graceful failure without dns server
This will fix the issues with download images when the dns setttings
are not correct.  On failure to download an image, it will now remove
the file that was created in prepration for download.  Due to ravens
simple checking for images based on names, that an image will only
exist when it has been correctly downloaded.